# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "SeedyFiuba API Gateway",
    "description": "API documentation for SeedyFiuba's API Gateway",
    "contact": {
      "name": "SeedyFiuba",
      "url": "https://github.com/Taller2SeedyFiuba"
    }
  },
...
  "paths": {
...
    "/projects/search": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Search for all projects matching given conditions.",
        "operationId": "searchProject",
        "parameters": [
          {
            "in": "query",
            "name": "ownerid",
            "schema": {
              "type": "string"
            },
            "description": "ID of the user who's projects want to be queried"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Type of the project to get"
          },
          {
            "in": "query",
            "name": "stage",
            "schema": {
              "type": "string"
            },
            "description": "Stage of the project to get"
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Tags of the project to get",
            "example": [
              "ArbolesParaChaco",
              "ChacoVive"
            ]
          },
          {
            "in": "query",
            "name": "lng",
            "schema": {
              "type": "number"
            },
            "description": "Longitude at which distance should be calculated"
          },
          {
            "in": "query",
            "name": "lat",
            "schema": {
              "type": "number"
            },
            "description": "Longitude at which distance should be calculated"
          },
          {
            "in": "query",
            "name": "dist",
            "schema": {
              "type": "number"
            },
            "description": "Distance from given latitude and longitud of the project to get in kilometers"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Maximum number of projects to retrieve"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number of projects result to retrieve"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeProject"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The specified query parameters are invalid."
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
    "/projects": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create project",
        "description": "This can only be done by the logged in user.",
        "operationId": "createProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ProjectBody"
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullProject"
                }
              }
            }
          },
          "400": {
            "description": "The specified project body information is not valid"
          },
          "404": {
            "description": "A user with de specified ID (owneruserid) was not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
    "/projects/{id}/sponsors": {
      "post": {
        "tags": [
          "sponsors"
        ],
        "summary": "Sponsor a new project specified by id",
        "operationId": "sponsorNewProject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the project to sponsor"
          }
        ],
        "requestBody": {
          "description": "A JSON object containing funding amount",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "integer",
                    "description": "Amount to provide",
                    "example": 20
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSponsor"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The specified project id is wrong."
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
    "/viewers": {
      "post": {
        "tags": [
          "viewers"
        ],
        "summary": "Add reviewing capabilities to the logged user",
        "operationId": "addReviewer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userid": {
                      "type": "string",
                      "example": "1lmoj2LKM10ADS"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "FullProject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 5
          },
          "ownerid": {
            "type": "string",
            "example": "05yseyhiEWPNvkYYbdHL77dHKWi1"
          },
          "title": {
            "type": "string",
            "example": "10000 Arboles para Chaco"
          },
          "description": {
            "type": "string",
            "example": "Este proyecto busca fondos para crear una obra de arte en Chaco"
          },
          "type": {
            "type": "string",
            "example": "arte"
          },
          "state": {
            "type": "string",
            "example": "funding"
          },
          "actualstage": {
            "type": "integer",
            "example": 0
          },
          "creationdate": {
            "type": "string",
            "example": "2021-03-14"
          },
          "location": {
            "type": "object",
            "properties": {
              "description": {
                "type": "string",
                "example": "Chaco, Argentina"
              },
              "lat": {
                "type": "number",
                "example": 120
              },
              "lng": {
                "type": "number",
                "example": 40
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "ArbolesParaChaco",
              "ChacoVerde",
              "ChacoVive"
            ]
          },
          "multimedia": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "https://picsum.photos/400/300",
              "https://picsum.photos/400/300",
              "https://picsum.photos/400/300"
            ]
          },
          "stages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "amount": {
                  "type": "integer"
                }
              }
            },
            "example": [
              {
                "title": "Arboles en el sur",
                "description": "Se plantaran 5000 arboles en el sur de Chaco",
                "amount": 2000
              },
              {
                "title": "Arboles en el norte",
                "description": "Se plantaran 5000 arboles en el norte de Chaco",
                "amount": 3000
              }
            ]
          },
          "totalamount": {
            "type": "integer",
            "example": 5000
          },
          "fundedamount": {
            "type": "integer",
            "example": 328
          },
          "sponsorscount": {
            "type": "integer",
            "example": 25
          },
          "favouritescount": {
            "type": "integer",
            "example": 10
          },
          "isfavourite": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "ResumeProject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 5
          },
          "ownerid": {
            "type": "string",
            "example": "05yseyhiEWPNvkYYbdHL77dHKWi1"
          },
          "title": {
            "type": "string",
            "example": "10000 Arboles para Chaco"
          },
          "type": {
            "type": "string",
            "example": "arte"
          },
          "state": {
            "type": "string",
            "example": "funding"
          },
          "location": {
            "type": "string",
            "example": "Chaco, Argentina"
          },
          "icon": {
            "type": "string",
            "example": "https://picsum.photos/400/300"
          }
        }
      },
      "ProjectSponsor": {
        "type": "object",
        "properties": {
          "userid": {
            "type": "string",
            "example": "05yseyhiEWPNvkYYbdHL77dHKWi1"
          },
          "projectid": {
            "type": "integer",
            "example": 5
          },
          "amount": {
            "type": "integer",
            "example": 18
          }
        }
      },
...
    },
    "requestBodies": {
...
      "ProjectBody": {
        "description": "A JSON object containing project information",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "10000 Arboles para Chaco"
                },
                "description": {
                  "type": "string",
                  "example": "Este proyecto busca fondos para crear una obra de arte en Chaco"
                },
                "type": {
                  "type": "string",
                  "example": "arte"
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "description": {
                      "type": "string",
                      "example": "Chaco, Argentina"
                    },
                    "lat": {
                      "type": "number",
                      "example": 120
                    },
                    "lng": {
                      "type": "number",
                      "example": 40
                    }
                  }
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "ArbolesParaChaco",
                    "ChacoVerde",
                    "ChacoVive"
                  ]
                },
                "multimedia": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "https://picsum.photos/400/300",
                    "https://picsum.photos/400/300",
                    "https://picsum.photos/400/300"
                  ]
                },
                "stages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "integer"
                      }
                    }
                  },
                  "example": [
                    {
                      "title": "Arboles en el sur",
                      "description": "Se plantaran 5000 arboles en el sur de Chaco",
                      "amount": 2000
                    },
                    {
                      "title": "Arboles en el norte",
                      "description": "Se plantaran 5000 arboles en el norte de Chaco",
                      "amount": 3000
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
...
}